+++ CHANGELOG.md
@@ -4,7 +4,8 @@
 
 ## fugit 1.0.0  not yet released
 
-* Provide Path.to_s
+* Differentiate `Dense.get(col, path)` from `Dense.fetch(col, path[, default])`
+* Provide Dense::Path.to_s
 * Introduce Dense.has_key?(collection, path)
 * Introduce Dense.insert(collection, path, value)
 * Accept `.first` and `.last` when indexing arrays